backend/requirements.txt
spotipy
lyricsgenius
python-dotenv
bs4
requests